requirements.txt
crewai[google-genai]>=0.28.0
langfuse>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-google-genai>=0.0.6
pydantic>=2.0.0
langchain-core>=0.1.0
